Rework form input binding
I forgot why I'm doing this, I guess it feels more correct idk.
Anyhow, I shifted from reading the properties from the DOM to using emits and storing the values in a ref.
Also I added more Typescript so I can more confidently use if/else and should help with validation later.

In the process I fixed the style of the periodicity switch, now it is always the same color and the selection text is highlighted